+++ apps/libcommon/proto.c
@@ -1,4 +1,4 @@
-// Copyright (C) 2022 - Tillitis AB
+// Copyright (C) 2022, 2023 - Tillitis AB
 // SPDX-License-Identifier: GPL-2.0-only
 
 #include <lib.h>
@@ -44,8 +44,8 @@ int parseframe(uint8_t b, struct frame_header *hdr)
 	case LEN_32:
 		hdr->len = 32;
 		break;
-	case LEN_128:
-		hdr->len = 128;
+	case LEN_512:
+		hdr->len = 512;
 		break;
 	default:
 		// Unknown length